상점 구현 완료 및 발사 및 재장전 재조정 시작

+++ Assets/Script/Weapon/Weapon.cs
@@ -21,7 +21,7 @@ public class Weapon : MonoBehaviour
      [SerializeField] public int magazine;
      [SerializeField] public int magazineCapacity;
      [SerializeField] public float firecooltime;
-     [SerializeField] public int reloadtime;
+     [SerializeField] public float reloadtime;
      [SerializeField] public GameObject dropPrefab;
      [SerializeField] public GameObject firepoint;
